Fixed compilation with current Agda-2.5.2.
Agda-2.5.2 requires deeper splitting in some cases.  Frankly, I do not
know why the code checks without this patch in Agda-2.5.1.1.  It seems
that without this deeper splitting, some defined types should not
reduce.

I often chose _ for variables which are not needed on the rhs.  You
might want to prettify this to match your style.