[ agda 2.6.2 compat ] Some recursive record need `eta-equality` now
These records are inferred as no-eta by Agda, which now leads to
rejection of matching against the constructor: agda/agda#4560.

The solution here is to declare these records as having eta-equality.

Alternatively, one could declare them as

  pattern
  no-eta-equality

but then this is no longer in the language of Agda 2.6.1.

In the long run (once Agda 2.6.2 is released), one should see whether
some speed-ups can be had with no-eta-equality.